pull implementation details of incremental compilation into its own file. rename some other files to be more descriptive
